+++ admin/maintain.go
@@ -15,7 +15,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package maintain
+package admin
 
 import (
 	"context"
